% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catHB.R
\name{catHB}
\alias{catHB}
\alias{print.catHB}
\title{Dynamic fit index (DFI) cutoffs for categorical multi-factor CFA models}
\usage{
catHB(
  model,
  n = NULL,
  plot = FALSE,
  manual = FALSE,
  reps = 250,
  estimator = "WLSMV",
  robust = FALSE
)

\method{print}{catHB}(x, ...)
}
\arguments{
\item{model}{This can either be a \code{\link{lavaan}} object from the \code{\link{cfa}} function,
OR a model statement written in \code{\link{lavaan}} \code{\link{model.syntax}} with standardized loadings and thresholds.}

\item{n}{If you entered a \code{\link{lavaan}} object for model, leave this blank.
Otherwise, enter your sample size (numeric).}

\item{plot}{Displays distributions of fit indices for each level of misspecification.}

\item{manual}{If you entered a \code{\link{lavaan}} object, keep this set to FALSE.
If you manually entered standardized loadings and sample size, set this to TRUE.}

\item{reps}{The number of replications used in your simulation. This is set to 500 by default in both the
R package and the corresponding Shiny App.}

\item{estimator}{Which estimator to use within the simulations (enter in quotes). The default is WLSMV. Only limited-information estimators that produce fit indices are permitted (i.e., maximum likelihood is not available).}

\item{robust}{A logical flag. When set to \code{TRUE}, the function will extract and report the \code{.robust}
fit indices (e.g., \code{rmsea.robust}, \code{cfi.robust}) from \code{lavaan} simulations. This is highly recommended
when using the \code{WLSMV} estimator to ensure the use of appropriate fit statistics corrected for non-normality,
as justified by the work of Brosseau-Liard, Savalei, & Li (2012) and Brosseau-Liard & Savalei (2014).
When \code{FALSE} (the default), the function retains its original behavior, which may report \code{.scaled} fit indices for certain estimators.}

\item{x}{catHB object}

\item{...}{other print parameters}
}
\value{
Dynamic fit index (DFI) cutoffs for SRMR, RMSEA, and CFI.
}
\description{
This function generates DFI cutoffs for multi-factor CFA models that treat items as categorical.
The default argument is a singular argument: a \code{\link{lavaan}} object from the \code{\link{cfa}} function.
The function can also accommodate manual entry of the model statement and sample size (including threshold estimates).

The app-based version of this function can be found at \href{https://dynamicfit.app/}{dynamicfit.app}.
}
\examples{
#Example using a lavaan object as input (manual=FALSE)

#two-factor model with correlated factors
m1<-"
 F1=~X1 + X2 +X3
 F2=~X6 + X7 + X8 + X9
 F1~~F2"

 #fit the model in lavaan, treating items are categorical
 fit<-lavaan::cfa(m1, data=Example, ordered=TRUE, estimator="WLSMV")

# Original behavior (robust=FALSE)
\donttest{catHB(fit)}

# Recommended behavior for WLSMV (robust=TRUE)
# This will produce different DFI cutoffs and empirical fit indices
\donttest{catHB(fit, robust=TRUE)}

#Manual entry example (manual=TRUE)

#two-factor model with correlated factors
m1<-"
 F1=~X1 + X2 +X3
 F2=~X6 + X7 + X8 + X12
 F1~~F2"

 #fit the model, treating items are categorical
 #lavaan is used here to show where estimates come from

 fit<-lavaan::cfa(m1, data=Example, ordered=TRUE)
 lavaan::standardizedsolution(fit)

manual_model <-"F1=~.448*X1 + .557*X2 + .770*X3
 F2=~.612*X6 + .684*X7 + .736*X8 + .365*X12
 F1~~.424*F2

X1 |-0.285*t1
X1 | 0.337*t2
X1 | 0.793*t3
X1 | 1.305*t4

X2 |-0.243*t1
X2 | 0.369*t2
X2 | 0.849*t3
X2 | 1.227*t4

X3 |-0.285*t1
X3 | 0.353*t2
X3 | 0.827*t3
X3 | 1.379*t4

X6 |-0.279*t1
X6 | 0.353*t2
X6 | 0.827*t3
X6 | 1.379*t4

X7 |-0.269*t1
X7 | 0.385*t2
X7 | 0.871*t3
X7 | 1.329*t4

X8 |-0.274*t1
X8 | 0.358*t2
X8 | 0.779*t3
X8 | 1.237*t4

X12 |-0.248*t1
X12 | 0.440*t2
X12 | 0.900*t3
X12 | 1.392*t4"
\donttest{catHB(model=manual_model,n=500,manual=TRUE)}

}
\references{
Brosseau-Liard, P. E., Savalei, V., & Li, L. (2012). An Investigation of the Sample Performance of Two Nonnormality Corrections for RMSEA. \emph{Multivariate Behavioral Research}, 47(6), 904-930.

Brosseau-Liard, P. E., & Savalei, V. (2014). Adjusting Incremental Fit Indices for Nonnormality. \emph{Multivariate Behavioral Research}, 49(5), 460-470.
}
\author{
Daniel McNeish & Melissa G Wolf

Maintainer: Daniel McNeish <dmcneish@asu.edu>
}
