% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfaFit.R
\name{cfaFit}
\alias{cfaFit}
\title{Deprecated: Fit index cutoffs for single level confirmatory factor analysis models}
\usage{
cfaFit(model, n)
}
\arguments{
\item{model}{Model statement written in \code{\link{lavaan}} \code{\link{model.syntax}} with standardized loadings.
If using \code{\link{lavaan}}, you can get standardized loadings from \code{\link{standardizedSolution}}.}

\item{n}{Sample size (number).}
}
\value{
Fit index cutoffs for the SRMR, RMSEA, and CFI.
}
\description{
Deprecated. If you want to replicate the cfaFit cutoffs for multi-factor models, use the function \code{\link{cfaHB}}.
The first level will be identical to the results from cfaFit. The cutoffs for one-factor models from cfaFit cannot be
replicated in other functions. Use \code{\link{cfaOne}} for one-factor CFA models.
}
\examples{
mod <- "F1 =~ .602*Y1 + .805*Y2 + .516*Y3 + .857*Y4
        F2 =~ .413*Y5 + -.631*Y6
        F1 ~~ .443*F2
        Y4 ~~ .301*Y5"
cfaFit(mod,500)
}
