% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDDFI.R
\name{DDDFI}
\alias{DDDFI}
\alias{print.DDDFI}
\title{Direct Discrepancy Dynamic fit index (3DFI) cutoffs for arbitrary covariance structure models}
\usage{
DDDFI(
  model,
  data = NULL,
  scale = "normal",
  manual = FALSE,
  reps = 250,
  n = NULL,
  estimator = NULL,
  MAD = c(0.038, 0.05, 0.06),
  plot.dfi = FALSE,
  plot.dist = FALSE,
  plot.discrepancy = FALSE
)

\method{print}{DDDFI}(x, ...)
}
\arguments{
\item{model}{This can either be a \code{\link{lavaan}} object, OR a model statement written in \code{\link{lavaan}} \code{\link{model.syntax}}
with standardized estimates}

\item{data}{The original data to which the model was applied. Not required if scale="normal". Otherwise, data is required.}

\item{scale}{Determines how data are simulated. Options are "normal", "nonnormal", or "categorical". "normal" assumes multivariate
normality across all variables. "nonnormal" recreates distributions in an empirical dataset (to be provided by the user), assuming variables are continuous.
"categorical" simulates discrete data with the same proportions as an empirical dataset (to be provided by the user). With "categorical", mixed formats
are also supported and any variable with more than 9 categories is simulated from a normal distribution.Only "normal" can be used without provided an original
dataset.}

\item{manual}{If you entered a \code{\link{lavaan}} object, keep this set to FALSE.
If you manually entered standardized estimates and sample size, set this to TRUE.}

\item{reps}{The number of replications used in the simulations. This is set to 250 by default}

\item{n}{If you entered a \code{\link{lavaan}} object for model, leave this blank.
Otherwise, enter your sample size (numeric).}

\item{estimator}{Which estimator to use within the simulations (enter in quotes). The default depends on the scale option ("ML" for "normal", "MLR" for "nonnormal",
and "WLSMV" for categorical)}

\item{MAD}{Mean Absolute Discrepancies to test in the simulation. Default is c(.038, .05, .06) to recreate traditional "Close", "Fair", "Mediocre" benchmarks}

\item{plot.dfi}{Displays simulated distributions of fit indices used to derive cutoffs for each MAD value.}

\item{plot.dist}{Displays distributions of simulated data (and empirical data, if provided) to assess fidelity of simulated data to empirical data}

\item{plot.discrepancy}{Displays distributions of simulated MAD values}

\item{x}{DDDFI object}

\item{...}{other print parameters}
}
\value{
Direct Discrepancy Dynamic fit index (DFI) cutoffs for CFI, RMSEA, and RMSEA 90% CI.
}
\description{
This function generates DFI cutoffs for any single group covariance structure model with
a saturated or absent mean structure. It supports (a) any estimator supported by lavaan (e.g.
ML, MLR, WLSMV, ULSMV), (b) missing data, and (c) multiple response scales (normal, non-normal continuous,
categorical). The default argument is a singular argument: a \code{\link{lavaan}} object.The function
can also accommodate manual entry of the model statement and sample size. Some features require an original
dataset to be provided (e.g., missing data, categorical data). The app-based version of
this function can be found at \href{https://dynamicfit.app/}{dynamicfit.app}.
}
\author{
Daniel McNeish & Melissa G Wolf

Maintainer: Daniel McNeish <dmcneish@asu.edu>
}
