% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example-data.R
\docType{data}
\name{Example}
\alias{Example}
\title{DFI Example Data}
\format{
## 'Example'
A data frame with 12 columns and 500 rows
\describe{
\item{X1}{Likert response to item 1}
\item{X2}{Likert response to item 2}
\item{X3}{Likert response to item 3}
\item{X4}{Likert response to item 4}
\item{X5}{Likert response to item 5}
\item{X6}{Likert response to item 6}
\item{X7}{Likert response to item 7}
\item{X8}{Likert response to item 8}
\item{X9}{Likert response to item 9}
\item{X10}{Likert response to item 10}
\item{X11}{Likert response to item 11}
\item{X12}{Likert response to item 12}
}
}
\source{
Simulated with simstandard package
}
\usage{
Example
}
\description{
Simulated 5-point Likert item responses to 12 items for 500 people.
}
\keyword{datasets}
