% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likertHB.R
\name{likertHB}
\alias{likertHB}
\alias{print.likertHB}
\title{Dynamic fit index (DFI) cutoffs multi-factor CFA models with likert items}
\usage{
likertHB(
  model,
  data,
  n = NULL,
  plot = FALSE,
  manual = FALSE,
  estimator = "ML",
  reps = 250
)

\method{print}{likertHB}(x, ...)
}
\arguments{
\item{model}{This can either be a \code{\link{lavaan}} object from the \code{\link{cfa}} function,
OR a model statement written in \code{\link{lavaan}} \code{\link{model.syntax}} with standardized loadings and thresholds.}

\item{data}{An empirical dataset used to determine the number of Likert responses and the response frequencies}

\item{n}{If you entered a \code{\link{lavaan}} object for model, leave this blank.
Otherwise, enter your sample size (numeric).}

\item{plot}{Displays distributions of fit indices for each level of misspecification.This also includes plots to visualize how close the
simulatied data are to the original data.}

\item{manual}{If you entered a \code{\link{lavaan}} object, keep this set to FALSE.
If you manually entered standardized loadings and sample size, set this to TRUE.}

\item{estimator}{Which estimator to use within the simulations (enter in quotes). The default is ML}

\item{reps}{The number of replications used in your simulation. This is set to 500 by default in both the
R package and the corresponding Shiny App.}

\item{x}{likertHB object}

\item{...}{other print parameters}
}
\value{
Dynamic fit index (DFI) cutoffs for SRMR, RMSEA, and CFI.
}
\description{
This function generates DFI cutoffs for multi-factor CFA models that treats Likert-type items as continuous.
 The default argument is a singular argument: a \code{\link{lavaan}} object from
 the \code{\link{cfa}} function. The function can also accommodate manual entry of the model statement and
 sample size. A primary difference in likert DFI functions is that a dataset must also be provided in the 'data' argument
 in order to simulate data with the same number of response options and response frequencies as the original data.

The app-based version of this function can be found at \href{https://dynamicfit.app/}{dynamicfit.app}.
}
\author{
Daniel McNeish & Melissa G Wolf

Maintainer: Daniel McNeish <dmcneish@asu.edu>
}
