% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likertHB2.R
\name{likertHB2}
\alias{likertHB2}
\alias{print.likertHB2}
\title{Dynamic fit index (DFI) cutoffs multi-factor CFA models with Likert-type items}
\usage{
likertHB2(
  model,
  data,
  n = NULL,
  plot = FALSE,
  manual = FALSE,
  estimator = "ML",
  reps = 250
)

\method{print}{likertHB2}(x, ...)
}
\arguments{
\item{model}{This can either be a \code{\link{lavaan}} object from the \code{\link{cfa}} function,
OR a model statement written in \code{\link{lavaan}} \code{\link{model.syntax}} with standardized loadings.}

\item{data}{An empirical dataset used to determine the number of Likert responses and the response frequencies}

\item{n}{If you entered a \code{\link{lavaan}} object for model, leave this blank.
Otherwise, enter your sample size (numeric).}

\item{plot}{Displays distributions of fit indices for each level of misspecification.
This also includes plots to visualize how close the simulated data are to the original data.}

\item{manual}{If you entered a \code{\link{lavaan}} object, keep this set to FALSE (the default).
If you manually entered standardized loadings and sample size, set this to TRUE.}

\item{estimator}{Which estimator to use within the simulations (enter in quotes). The default is ML}

\item{reps}{The number of replications used in your simulation. This is set to 500 by default.}

\item{x}{likertHB2 object}

\item{...}{other print parameters}
}
\value{
Dynamic fit index (DFI) cutoffs for SRMR, RMSEA, and CFI.
}
\description{
This function generates DFI cutoffs for multi-factor CFA models that treats Likert-type items as continuous.
 The default argument is a singular argument: a \code{\link{lavaan}} object from
 the \code{\link{cfa}} function. The function can also accommodate manual entry of the model statement and
 sample size. A primary difference in likert DFI functions is that a dataset must also be provided in the 'data' argument
 in order to simulate data with the same number of response options and response frequencies as the original data.
 When Likert-type items are treated as continuous, this information cannot be obtained solely from model output.

The app-based version of this function can be found at \href{https://dynamicfit.app/}{dynamicfit.app}.
}
\examples{
#Example using a lavaan object as input (manual=FALSE)

#two-factor model with correlated factors
m1<-"
 F1=~X2 + X3 + X4
 F2=~X6 + X7 + X8
 F1~~F2"

 #fit the model in lavaan, treating items are continuous
 fit<-lavaan::cfa(m1, data=Example)

\donttest{likertHB2(fit, data=Example)}

#Manual entry example (manual=TRUE)

#two-factor model with correlated factors
m1<-"
 F1=~X2 + X3 + X4
 F2=~X6 + X7 + X8
 F1~~F2"

 #fit the model, treating items are continuous
 #lavaan is used here to shown where estimates come from
 #but manual entry supports standardized estimates from models fit in any software

 fit<-lavaan::cfa(m1, data=Example)
 lavaan::standardizedsolution(fit)


manual_model <-"F1=~.554*X2 + .654*X3 + .733*X4
 F2=~.537*X6 + .666*X7 + .723*X8
 F1~~.339*F2"

\donttest{likertHB2(model=manual_model,data=Example,n=500,manual=TRUE)}

}
\author{
Daniel McNeish & Melissa G Wolf

Maintainer: Daniel McNeish <dmcneish@asu.edu>
}
