% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnorOne.R
\name{nnorOne}
\alias{nnorOne}
\alias{print.nnorOne}
\title{Dynamic fit index (DFI) cutoffs for continuous, non-normal one-factor CFA models with (possible) missing data}
\usage{
nnorOne(
  model,
  data,
  n = NULL,
  plot = FALSE,
  manual = FALSE,
  estimator = "MLR",
  reps = 500
)

\method{print}{nnorOne}(x, ...)
}
\arguments{
\item{model}{This can either be a \code{\link{lavaan}} object from the \code{\link{cfa}} function,
OR a model statement written in \code{\link{lavaan}} \code{\link{model.syntax}} with standardized loadings}

\item{data}{An empirical dataset to which a modified Bollen-Stine bootstrap will be applied to create hypothetical misspecified data}

\item{n}{If you entered a \code{\link{lavaan}} object for model, leave this blank.
Otherwise, enter your sample size (numeric).}

\item{plot}{Displays distributions of fit indices for each level of misspecification.This also includes plots to visualize how close the
distributions of the hypothetical data come to the original data.}

\item{manual}{If you entered a \code{\link{lavaan}} object, keep this set to FALSE.
If you manually entered standardized loadings and sample size, set this to TRUE.}

\item{estimator}{Which estimator to use within the simulations (enter in quotes). The default is MLR}

\item{reps}{The number of replications used in your simulation. This is set to 500 by default in both the
R package and the corresponding Shiny App.}

\item{x}{nnorOne object}

\item{...}{other print parameters}
}
\value{
Dynamic fit index (DFI) cutoffs for SRMR, RMSEA, and CFI.
}
\description{
This function generates DFI cutoffs for one-factor CFA models that treats items as continuous and non-normal with
possible missing data. This functions uses a modified Bollen-Stine bootstrap to accommodate non-normality and missingness rather
than simulating from a particular distribution. The default argument is a singular argument: a \code{\link{lavaan}} object from
 the \code{\link{cfa}} function. The function can also accommodate manual entry of the model statement and
 sample size (including threshold estimates). A primary difference in nnor DFI functions is that a dataset from which to bootstrap
 must also be provided in the 'data' argument.

The app-based version of this function can be found at \href{https://dynamicfit.app/}{dynamicfit.app}.
}
\examples{
#Example using a lavaan object as input (manual=FALSE)

#one-factor model
m1<-"F1=~X5+ X6 + X7 + X8 + X9"

 #fit the model in lavaan, treating items are continuous
 fit<-lavaan::cfa(m1, data=Example)

\donttest{nnorOne(fit, data=Example)}

#Manual entry example (manual=TRUE)

#one-factor model with correlated factors
m1<-"F1=~X5+ X6 + X7 + X8 + X9"

 #fit the model, treating items are continuous
 #lavaan is used here to shown where estimates come from
 #but manual entry supports standardized estimates from models fit in any software

 fit<-lavaan::cfa(m1, data=Example)
 lavaan::standardizedsolution(fit)

manual_model <-"F1=~.517*X5 + .549*X6 + .679*X7 + .694*X8 + .203*X9"

\donttest{nnorOne(model=manual_model,data=Example,n=500,manual=TRUE)}

}
\author{
Daniel McNeish & Melissa G Wolf

Maintainer: Daniel McNeish <dmcneish@asu.edu>
}
