% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelRep.R
\name{RelRep}
\alias{RelRep}
\title{Reliability Representativeness of Coefficient Alpha or Omega}
\usage{
RelRep(
  data,
  items = c(names(data)),
  rel = "alpha",
  missing = "NA",
  method = "CI",
  width = NULL,
  raw.low = NULL,
  raw.high = NULL
)
}
\arguments{
\item{data}{The original data to which the model was applied.}

\item{items}{Column names of the items on the scale being evaluated (entered as strings). If omitted, all variables in the data will be used.}

\item{rel}{Reliability coefficient to analyze. Options are "alpha" (the default) or "omega".}

\item{missing}{The missing data indicator in the data. Not needed in R, only present to simply use of this function in a Shiny application.}

\item{method}{how the test interval is created. Options are "CI" (the default), "width", or "raw".
"CI" uses a 95% Bayesian highest posterior density credible interval.
"width" builds an interval using a predetermined relative distance from the reliability coefficient (e.g., .05 from alpha).
"raw" builds an interval using a predetermined raw values (e.g., .70 to .90)}

\item{width}{Only required if method="width". Specifies a predetermined relative distance from the coefficient to each bound of the interval.
The total width of the interval will be twice this value
(e.g., if .05 is entered, the total interval width is .10 because it will span .05 above the coefficient and .05 below the coefficient)}

\item{raw.low}{Only required if method="raw". Manually specifies the lower bound of the test interval. Must be between 0 and 1.}

\item{raw.high}{Only required if method="raw". Manually specifies the upper bound of the test interval. Must be between 0 and 1.}
}
\value{
Conditional reliability table and Reliability Representativeness plot and table.
}
\description{
This function evaluates how well a reliability summary index like alpha or omega
 represents the conditional reliability of a distribution of composite scores. It compares the conditional
 reliability function to a summary index and outputs a representativeness plot, a table of representativeness indices,and
 the full conditional reliability table for each possible sum score.
}
\examples{

# "Example" dataset has 12 items on a 1-5 Likert scale

#Example using the first 8 items in "Example" for testing coefficient alpha with a 95\% Bayes credible interval
ex1<-RelRep(data=Example, items=c(names(Example[,1:8]), rel="alpha", method="CI"))

#Example using odd items in "Example" to build a test interval that is .05 above and below coefficient omega
ex2<-RelRep(data=Example, items=c("X1","X3","X5","X7","X9","X11"), rel="omega", method="width", width=.05)

#Example using even items in "Example" to specify a test interval for coefficient alpha between 0.70 and 0.80
ex3<-RelRep(data=Example, items=c("X2","X4","X6","X8","X10","X12"), rel="alpha", method="raw", raw.low=.70, raw.high=.80)
}
\author{
Daniel McNeish & Denis Dumas

Maintainer: Daniel McNeish <dmcneish@asu.edu>
}
