% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catHB.R
\name{catHB}
\alias{catHB}
\alias{print.catHB}
\title{Dynamic fit index (DFI) cutoffs for categorical multi-factor CFA models}
\usage{
catHB(
  model,
  n = NULL,
  plot = FALSE,
  manual = FALSE,
  reps = 250,
  estimator = "WLSMV"
)

\method{print}{catHB}(x, ...)
}
\arguments{
\item{model}{This can either be a \code{\link{lavaan}} object from the \code{\link{cfa}} function,
OR a model statement written in \code{\link{lavaan}} \code{\link{model.syntax}} with standardized loadings and thresholds.}

\item{n}{If you entered a \code{\link{lavaan}} object for model, leave this blank.
Otherwise, enter your sample size (numeric).}

\item{plot}{Displays distributions of fit indices for each level of misspecification.}

\item{manual}{If you entered a \code{\link{lavaan}} object, keep this set to FALSE.
If you manually entered standardized loadings and sample size, set this to TRUE.}

\item{reps}{The number of replications used in your simulation. This is set to 500 by default in both the
R package and the corresponding Shiny App.}

\item{estimator}{Which estimator to use within the simulations (enter in quotes). The default is WLSMV. Only limited-information estimators that produce fit indices are permitted (i.e., maximum likelihood is not available)}

\item{x}{catHB object}

\item{...}{other print parameters}
}
\value{
Dynamic fit index (DFI) cutoffs for SRMR, RMSEA, and CFI.
}
\description{
This function generates DFI cutoffs for multi-factor CFA models that treat items as categorical.
The default argument is a singular argument: a \code{\link{lavaan}} object from the \code{\link{cfa}} function.
The function can also accommodate manual entry of the model statement and sample size (including threshold estimates).

The app-based version of this function can be found at \href{https://dynamicfit.app/}{dynamicfit.app}.
}
\examples{
#Lavaan object example (manual=FALSE)
dat <- lavaan::HolzingerSwineford1939
lavmod <- "F1 =~ x1 + x2 + x3 + x4 + x5 + x6 + x7 + x8 + x9"
fit <- lavaan::cfa(lavmod,dat)
\donttest{cfaOne(fit)}

#Manual entry example for a sample size of 6649 (manual=TRUE)
#thresholds go in model statement as (a)categorical item name, (b) vertical pipe, (c) estimate, (d)times t+threshold number
manmod <- "E1 =~ .76*bfi_e1 + .73*bfi_e2 + .59*bfi_e3
E2 =~ .71*bfi_e4 + .84*bfi_e5 + .58*bfi_e6
E3 =~ .71*bfi_e7 + .80*bfi_e8 + .61*bfi_e9

E1 ~~ .25*E2
E1 ~~ .35*E3
E2 ~~ .45*E3

bfi_e1 |-1.69*t1
bfi_e1 |-1.06*t2
bfi_e1 |-0.53* t3
bfi_e1 |0.06* t4
bfi_e1 |0.75* t5

bfi_e2 |-1.16*t1
bfi_e2 |-0.42*t2
bfi_e2 |0.28*t3
bfi_e2 |0.71*t4
bfi_e2 |1.34*t5

bfi_e3 |-1.99* t1
bfi_e3 |-1.27* t2
bfi_e3 | 0.61*t3
bfi_e3 | 0.09*t4
bfi_e3 | 0.97*t5

bfi_e4 |-2.05* t1
bfi_e4 | -1.36*t2
bfi_e4 | -0.74*t3
bfi_e4 | -.004*t4
bfi_e4 | 0.81*t5

bfi_e5 | -1.12*t1
bfi_e5 | -0.42*t2
bfi_e5 | 0.16*t3
bfi_e5 | 0.60*t4
bfi_e5 | 1.21*t5

bfi_e6 | -1.76*t1
bfi_e6 | -1.18*t2
bfi_e6 | -0.68*t3
bfi_e6 | -0.03*t4
bfi_e6 | 0.76*t5

bfi_e7 | -1.04*t1
bfi_e7 | -0.31*t2
bfi_e7 | 0.46*t3
bfi_e7 | 0.78*t4
bfi_e7 | 1.38*t5

bfi_e8 | -1.74*t1
bfi_e8 | -1.12*t2
bfi_e8 | -0.64*t3
bfi_e8 | -.002*t4
bfi_e8 | 0.77*t5

bfi_e9 | -1.74*t1
bfi_e9 | -1.12*t2
bfi_e9 | -0.64*t3
bfi_e9 | -.002*t4
bfi_e9 | 0.77*t5"
\donttest{catHB(model=manmod,n=6649,manual=TRUE)}

}
\author{
Daniel McNeish & Melissa G Wolf

Maintainer: Daniel McNeish <dmcneish@asu.edu>
}
