% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfaFit.R
\name{cfaFit}
\alias{cfaFit}
\title{Fit index cutoffs for single level confirmatory factor analysis models}
\usage{
cfaFit(model, n)
}
\arguments{
\item{model}{Model statement written in \code{\link{lavaan}} \code{\link{model.syntax}} with standardized loadings.
If using \code{\link{lavaan}}, you can get standardized loadings from \code{\link{standardizedSolution}}.}

\item{n}{Sample size (number).}
}
\value{
Fit index cutoffs for the SRMR, RMSEA, and CFI.
}
\description{
Use this command to generate fit index cutoffs that are tailored to your CFA model and sample size.
}
\examples{
mod <- "F1 =~ .602*Y1 + .805*Y2 + .516*Y3 + .857*Y4
        F2 =~ .413*Y5 + -.631*Y6
        F1 ~~ .443*F2
        Y4 ~~ .301*Y5"
cfaFit(mod,500)
}
