% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDDFI.R
\name{cleanmodel}
\alias{cleanmodel}
\alias{print.DDDFI}
\title{Direct Discrepancy Dynamic fit index (3DFI) cutoffs for arbitrary covariance structure models}
\usage{
cleanmodel(model)

\method{print}{DDDFI}(x, ...)
}
\arguments{
\item{model}{This can either be a \code{\link{lavaan}} object, OR a model statement written in \code{\link{lavaan}} \code{\link{model.syntax}}
with standardized estimates}

\item{x}{DDDFI object}

\item{...}{other print parameters}

\item{scale}{Determines how data are simulated in the function. Options are "normal", "nonnormal", or "categorical". "normal" assumes multivariate
normality across all variables. "nonnormal" recreates distributions in an empirical dataset (to be provided by the user), assuming variables are continuous.
"categorical" simulates discrete data with the same proportions as an empirical dataset (to be provided by the user). With "categorical", mixed formats
are also supported and any variable with more than 9 categories is simulated from a normal distribution.}

\item{data}{The original data to which the model was applied. Not required if scale="normal". Otherwise, data is required.}

\item{manual}{If you entered a \code{\link{lavaan}} object, keep this set to FALSE.
If you manually entered standardized estimates and sample size, set this to TRUE.}

\item{reps}{The number of replications used in your simulation. This is set to 250 by default}

\item{n}{If you entered a \code{\link{lavaan}} object for model, leave this blank.
Otherwise, enter your sample size (numeric).}

\item{estimator}{Which estimator to use within the simulations (enter in quotes). The default is "ML".}

\item{MAD}{Mean Absolute Discrepancies to test in the simulation. Default is c(.038, .05, .06) to recreate traditional "Close", "Fair", "Mediocre" benchmarks}

\item{plot.dfi}{Displays simulated distributions of fit indices used to derive cutoffs for each level of misspecification.}

\item{plot.dist}{Displays distributions of simulated data (and empirical data, if provided) to assess fidelity of simulated data to empirical data}

\item{plot.discrepancy}{Displays distributions of simulated MAD values}
}
\value{
Direct Discrepancy Dynamic fit index (DFI) cutoffs for CFI, RMSEA, and RMSEA 90% CI.
}
\description{
This function generates DFI cutoffs for any single group covariance structure model with
no mean structure or a saturated mean structure. It supports (a) any estimator supported by lavaan (e.g.
ML, MLR, WLSMV, ULSMV), (b) missing data, and (c) multiple data formats (normal, non-normal continuous,
categorical). The default argument is a singular argument: a \code{\link{lavaan}} object.The function
can also accommodate manual entry of the model statement and sample size. Some features require an original
dataset to be provided as well (e.g., missing data, categorical data). The app-based version of
this function can be found at \href{https://dynamicfit.app/}{dynamicfit.app}.
}
\author{
Daniel McNeish & Melissa G Wolf

Maintainer: Daniel McNeish <dmcneish@asu.edu>
}
